Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -37,7 +38,7 @@ class QuestionsController extends Controller
     {
         $request->user()->questions()->create($request->only('title', 'body'));
 
-        return redirect()->route('questions.index')->with('success', "Your question has been submitted.");
+        return redirect()->route('questions.index')->with('success', 'Your question has been submitted.');
     }
 
     /**
@@ -57,7 +58,11 @@ class QuestionsController extends Controller
     {
         $question = Question::findOrFail($id);
 
-        return view("questions.edit", compact('question'));
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
+        return view('questions.edit', compact('question'));
     }
 
     /**
@@ -65,9 +70,13 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->update($request->only('title', 'body'));
 
-        return redirect('/questions')->with('success', "Your question has been updated.");
+        return redirect('/questions')->with('success', 'Your question has been updated.');
     }
 
     /**
@@ -75,8 +84,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->delete();
 
-        return redirect('/questions')->with('success', "Your question has been deleted.");
+        return redirect('/questions')->with('success', 'Your question has been deleted.');
     }
 }
